Use Lembrame as a python module
